Reworked command line code
