+++ image_titler/command.py
@@ -4,7 +4,7 @@ from tkinter.filedialog import askopenfilename, askdirectory
 from typing import Optional
 
 from image_titler.utilities import process_batch, process_image, TIER_MAP, FILE_TYPES, save_copy, \
-    convert_file_name_to_title, FONT
+    convert_file_name_to_title, FONT, parse_input
 
 
 def _request_input_path(path: str, batch: bool) -> Optional[str]:
@@ -32,26 +32,7 @@ def _request_input_path(path: str, batch: bool) -> Optional[str]:
     return input_path
 
 
-def parse_input() -> argparse.Namespace:
-    """
-    Creates and executes a parser on the command line inputs.
-
-    :return: the processed command line arguments
-    """
-    parser = argparse.ArgumentParser()
-    parser.add_argument('-t', '--title', help="add a custom title to the image (no effect when batch processing)")
-    parser.add_argument('-p', '--path', help="select an image file")
-    parser.add_argument('-o', '--output_path', help="select an output path for the processed image")
-    parser.add_argument('-r', '--tier', default="", choices=TIER_MAP.keys(),
-                        help="select an image tier")
-    parser.add_argument('-l', '--logo_path', help="select a logo file for addition to the processed image")
-    parser.add_argument('-b', '--batch', default=False, action='store_true', help="turn on batch processing")
-    parser.add_argument('-f', "--font", default=FONT, help="add a custom font by path (e.g. 'arial.ttf')")
-    args = parser.parse_args()
-    return args
-
-
-def title_image(args: argparse.Namespace) -> None:
+def _title_image(args: argparse.Namespace) -> None:
     """
     Titles an image based on a set of arguments.
 
@@ -95,7 +76,7 @@ def main() -> None:
     :return: None
     """
     args = parse_input()
-    title_image(args)
+    _title_image(args)
 
 
 if __name__ == '__main__':
